refactor(store): extract filter slice + introduce slice infrastructure
- Move WorkspaceState + UndoState + MAX_UNDO into src/store/workspace-types.ts
  so per-domain slices can import the type without circling back through
  workspace.ts.
- Create src/store/slices/filter-slice.ts containing the highlighter filter
  state (active*Filter, *FilterMode) and all setters/togglers.
- Compose via createFilterSlice in the immer create() call.

workspace.ts: 1216 → 990 LOC. First slice as a proof-of-concept for the
TS plumbing; addresses the Simplicity audit MAJOR (split workspace.ts
into per-domain slices) incrementally.

966/966 vitest passing (969 with codex's recently added tests).
